pyproject.toml

[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "pydantic>=2.5.0",
    "starlette>=0.27.0",
    "google-genai>=1.27.0",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "langchain>=0.3.27",
    "pypdf>=5.9.0",
    "itsdangerous>=2.2.0",
    "simsimd>=6.5.0",
]
